+++ src/appleseed/renderer/kernel/rendering/generic/generictilerenderer.cpp
@@ -121,7 +121,7 @@ namespace
             const Frame&    frame,
             const size_t    tile_x,
             const size_t    tile_y,
-            const size_t    pass_hash,
+            const uint32    pass_hash,
             IAbortSwitch&   abort_switch) override
         {
             // Retrieve frame properties.
